+++ source/pyreshaper/test/s2sReshaperTests.py
@@ -1,7 +1,7 @@
 """
 Parallel Tests for the Reshaper class
 
-Copyright 2015, University Corporation for Atmospheric Research
+Copyright 2016, University Corporation for Atmospheric Research
 See the LICENSE.rst file for details
 """