+++ CHANGES.rst
@@ -1,7 +1,7 @@
 PyReshaper Change Log
 =====================
 
-Copyright 2015, University Corporation for Atmospheric Research
+Copyright 2016, University Corporation for Atmospheric Research
 See the LICENSE.rst file for details
 
 VERSION 0.9.0