+++ CHANGES.rst
@@ -132,4 +132,4 @@ VERSION 1.0.0
    removing hierarchy of matching Specifier classes)
  - No longer requires PyNIO to install...but you need netCDF4 or PyNIO to
    run the reshaper!
- - 
+